extracted search result into separate model
